+++ packages/creem-sdk/RELEASES.md
@@ -59,3 +59,13 @@ Based on:
 - [typescript v1.5.1] packages/creem-sdk
 ### Releases
 - [NPM v1.5.1] https://www.npmjs.com/package/creem/v/1.5.1 - packages/creem-sdk
+
+## 2026-06-22 11:26:00
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.784.0 (2.911.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v1.5.2] packages/creem-sdk
+### Releases
+- [NPM v1.5.2] https://www.npmjs.com/package/creem/v/1.5.2 - packages/creem-sdk